move shop to shop.ts
